+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
-- Feature: Add new `Checkbox` React component.
 
+- Feature: Add new `Checkbox` React component.
+- Feature: Add new `FormActions` React component.
 - Feature: Add new `Grid` and `GridCol` React components.
 - Fix: At xxs, use thinner gutters.